move functions about data path from ofUtils to ofFileUtils
